requirements.txt
django==4.0.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
psycopg2==2.9.1
gunicorn==20.1.0
django-heroku==0.3.1

pyJWT==2.1.0
asgiref==3.4.1
pytz==2021.1
sqlparse==0.4.2
whitenoise==5.3.0